Update question1.pl
Not sure if this will add it to a new branch/file or not...this is the first time I've forked someone else's project, so I've added my solution to the bottom of yours. 

I actually thought of sorting the list in ascending order as you've done, but I read iterate in the tute question and figured I'd just use a plain old for loop.